+++ Perceptron/funciones.h
@@ -27,5 +27,5 @@ list <string>  construirNgrama(string, int);
 bool buscarStopWord(string word,list<string> & listaStopWords);
 
 //Incrementa en uno en la posicion indicada;
-void incrementar(list<unsigned int> & hashTable,unsigned long int posicion );
+void incrementar(list<unsigned short int> & hashTable,unsigned long int posicion );
